requirements-dev.txt
pytest-django
selenium
pylint
pylint-django
black
